Add files via upload
I'm not sure if this description will save but here I go again. FIngers crossed! In practice4 I was using the code I had completed yesterday about computing pay and added a try/except in order to filter bad responses that were not integers so the code wouldn't break if someone didn't enter the right input. Pretty straightforward, but definitely going to need some more practice to become familiar with the indenting. Copilot is helping a ton when I finish to clean up my code, but I'd like to eventually get to the point where I'm mostly doing things on my own. 

Next, in practice5 I used the same code again but added to it and made part of it into a function, then called it to produce the same results. I added some things to make it look a little cleaner and played around a little again trying not to just do exactly what the instructor is doing in his videos. I did some trial and error and took a little longer so I only completed two exercises today but I did gain some good knowledge tinkering with my code. Things are definitley ramping up. I needed my MudWater early today. It's hard but definitely satisfying when you execute the program and it actually does what you want it to do. Till tomorrow, signing off :)